December 2016 revision
See the updated preface for a list of changes. A number of errors
encountered during use for the Fall 2016 semester have been corrected.
The polar coordinates section has been streamlined, and new examples and
exercises have been added in Sections 4.2 and 4.5.
A new section on Diagonalization has also been created.